dbt/pyproject.toml
[project]
name = "dbt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-core>=1.11.2",
    "dbt-snowflake==1.9.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "snowflake-connector-python[pandas]>=3.18.0",
    "streamlit>=1.52.2",
]
